refactor(granite): map token embeddings by row
